+++ tests/ComponentRegistrationTest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Btekno\Mail\Tests;
+namespace Nuewire\Mail\Tests;
 
 use Livewire\Livewire;
 
@@ -10,7 +10,7 @@ final class ComponentRegistrationTest extends TestCase
 {
     public function test_the_settings_component_is_registered(): void
     {
-        Livewire::test('btekno::mail')
+        Livewire::test('nuewire::mail')
             ->assertSee('Pengaturan Email');
     }
 }